PR review comments: moved NodeInfoFilesCopier under node-api
