+++ node-api/src/test/kotlin/net/corda/nodeapi/NodeInfoFilesCopierTest.kt
@@ -1,9 +1,9 @@
-package net.corda.utilities
+package net.corda.nodeapi
 
 import net.corda.cordform.CordformNode
 import net.corda.core.identity.CordaX500Name
 import net.corda.core.utilities.NetworkHostAndPort
-import net.corda.core.utilities.NodeInfoFilesCopier
+import net.corda.node.services.network.NODE_INFO_FILE_NAME_PREFIX
 import net.corda.testing.eventually
 import org.junit.Before
 import org.junit.Rule
@@ -31,8 +31,8 @@ class NodeInfoFilesCopierTest {
         private const val NODE_2_PATH = "node2"
 
         private val content = "blah".toByteArray(Charsets.UTF_8)
-        private val GOOD_NODE_INFO_NAME = "nodeInfo-test"
-        private val GOOD_NODE_INFO_NAME_2 = "nodeInfo-anotherNode"
+        private val GOOD_NODE_INFO_NAME = "${NODE_INFO_FILE_NAME_PREFIX}test"
+        private val GOOD_NODE_INFO_NAME_2 = "${NODE_INFO_FILE_NAME_PREFIX}anotherNode"
         private val BAD_NODE_INFO_NAME = "something"
         private val legalName = CordaX500Name(organisation = ORGANIZATION, locality = "Nowhere", country = "GB")
         private val hostAndPort = NetworkHostAndPort("localhost", 1)
